Record app performance improvements
The Record app now uses a separate thread for writing packet data to the recording file. This allows the proxy to continue forwarding packets without having to wait for the data to be written to the underlying stream. Having a separate thread also allows the app to wait for all data to be written to disk without relying on a Sleep() call.