[12.15.8493] Tournaments protocol (#24)
* Add Logger class

Adds Logger class to Utilities and makes use of it in the API, and allows the level and output to be set via the command line for the Extract, Record, and Watch apps.

* Cleanup Communication class

Reduces the insane amount of repetitive code in the Communication class

* Change message parsing override method from bool to void

* Add new tournament-related packets

* Remove unused override methods

* Address additional unknowns for character info and tournaments

* Address most unknowns in TournamentInformation server packet

* Prepare for new, unknown packets

* Update protocol related to purchasing Tournament ticket from Store

* Record app performance improvements

The Record app now uses a separate thread for writing packet data to the recording file. This allows the proxy to continue forwarding packets without having to wait for the data to be written to the underlying stream. Having a separate thread also allows the app to wait for all data to be written to disk without relying on a Sleep() call.

* Address unknown client packets related to Tournaments

* Move logging to a separate thread
Like with the Record app improvements, this will move logging to a
separate thread to reduce time it takes for the calling code to return.