Enable appending of most packets with unknowns
There are still a lot of unknowns throughout both client and server packets. However, some of these unknowns are simple enough that they can still be used to append a packet to a NetworkMessage.

This commit also fixes the RSA start position in the Watch app, and includes a various few other tweaks.